# mcd_care_admin/src/generated/api/models/account-response.ts
/* tslint:disable */
/* eslint-disable */
export interface AccountResponse {
  created?: string;
  email?: null | string;
  firstName?: null | string;
  id?: null | string;
  isVerified?: boolean;
  lastName?: null | string;
  role?: null | string;
  updated?: null | string;
  username?: null | string;
}