# mcd_care_admin/src/generated/api/models/authenticate-response.ts
/* tslint:disable */
/* eslint-disable */
export interface AuthenticateResponse {
  created?: string;
  email?: null | string;
  firstName?: null | string;
  id?: null | string;
  isVerified?: boolean;
  jwtToken?: null | string;
  lastName?: null | string;
  role?: null | string;
  updated?: null | string;
}