# mcd_care_admin/src/generated/api/models/create-request.ts
/* tslint:disable */
/* eslint-disable */
export interface CreateRequest {
  confirmPassword: string;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  role: string;
  username: string;
}