# mcd_care_admin/src/generated/api/models/register-request.ts
/* tslint:disable */
/* eslint-disable */
export interface RegisterRequest {
  acceptTerms?: boolean;
  confirmPassword: string;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  username: string;
}