# mcd_care_admin/src/generated/api/models/update-request.ts
/* tslint:disable */
/* eslint-disable */
export interface UpdateRequest {
  confirmPassword?: null | string;
  email?: null | string;
  firstName?: null | string;
  lastName?: null | string;
  password?: null | string;
  role?: null | string;
}